scripts/requirements.txt
opencv-python
Pillow
pyserial
av
pySerialTransfer
pygame